+++ samples/WeihanLi.Web.Extensions.Samples/ValuesController.cs
@@ -1,8 +1,10 @@
 ﻿// Copyright (c) Weihan Li. All rights reserved.
 // Licensed under the MIT license.
 
+using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.ComponentModel.DataAnnotations;
 using System.Security.Claims;
 using WeihanLi.Common.Models;
 using WeihanLi.Common.Services;
@@ -10,20 +12,38 @@ using WeihanLi.Web.Authentication.ApiKeyAuthentication;
 using WeihanLi.Web.Authentication.HeaderAuthentication;
 using WeihanLi.Web.Authentication.QueryAuthentication;
 using WeihanLi.Web.Authorization.Token;
-using WeihanLi.Web.Middleware;
+using WeihanLi.Web.Filters;
 
 namespace WeihanLi.Web.Extensions.Samples;
 
 [Route("api/values")]
 [ApiController]
+[ApiResultFilter]
 public class ValuesController : ControllerBase
 {
+    private readonly IServiceScopeFactory _serviceScopeFactory;
+
+    public ValuesController(IServiceScopeFactory serviceScopeFactory) => _serviceScopeFactory = serviceScopeFactory;
+
+    [HttpGet("[action]")]
+    public IActionResult ServiceScopeTest()
+    {
+        Task.Run(() =>
+        {
+            using var scope = _serviceScopeFactory.CreateScope();
+            var tokenService = scope.ServiceProvider.GetRequiredService<ITokenService>();
+            Console.WriteLine(tokenService.GetHashCode());
+        });
+        return Ok();
+    }
+
     [HttpGet]
     public async Task<IActionResult> Get([FromServices] IUserIdProvider userIdProvider)
     {
         var headerAuthResult = await HttpContext.AuthenticateAsync(HeaderAuthenticationDefaults.AuthenticationSchema);
         var queryAuthResult = await HttpContext.AuthenticateAsync(QueryAuthenticationDefaults.AuthenticationSchema);
         var apiKeyAuthResult = await HttpContext.AuthenticateAsync(ApiKeyAuthenticationDefaults.AuthenticationSchema);
+        var bearerAuthResult = await HttpContext.AuthenticateAsync(JwtBearerDefaults.AuthenticationScheme);
 
         return Ok(new
         {
@@ -32,6 +52,7 @@ public class ValuesController : ControllerBase
             headerAuthResult = headerAuthResult.Principal?.Identity,
             queryAuthResult = queryAuthResult.Principal?.Identity,
             apiKeyAuthResult = apiKeyAuthResult.Principal?.Identity,
+            bearerAuthResult = bearerAuthResult.Principal?.Identity
         });
     }
 
@@ -63,13 +84,15 @@ public class ValuesController : ControllerBase
     }
 
     [HttpGet("getToken")]
-    public async Task<IActionResult> GetToken(string userName, [FromServices] ITokenService tokenService)
+    public async Task<IActionResult> GetToken([Required] string userName, [FromServices] ITokenService tokenService)
     {
-        return await tokenService
-            .GenerateToken(new Claim("name", userName))
-            .ContinueWith(r =>
-                r.Result.WrapResult().GetRestResult()
-            );
+        var token = await tokenService
+            .GenerateToken(new Claim("name", userName));
+        if (token is TokenEntityWithRefreshToken tokenEntityWithRefreshToken)
+        {
+            return tokenEntityWithRefreshToken.WrapResult().GetRestResult();
+        }
+        return token.WrapResult().GetRestResult();
     }
 
     [HttpGet("validateToken")]
@@ -80,5 +103,30 @@ public class ValuesController : ControllerBase
             .ContinueWith(r =>
                 r.Result.WrapResult().GetRestResult()
             );
+    }
+
+    [HttpGet("RefreshToken")]
+    public async Task<IActionResult> RefreshToken(string refreshToken, [FromServices] ITokenService tokenService)
+    {
+        var token = await tokenService
+            .RefreshToken(refreshToken);
+        if (token is TokenEntityWithRefreshToken tokenEntityWithRefreshToken)
+        {
+            return tokenEntityWithRefreshToken.WrapResult().GetRestResult();
+        }
+        return token.WrapResult().GetRestResult();
+    }
+
+    [HttpGet("[action]")]
+    [Authorize(AuthenticationSchemes = "Bearer")]
+    public IActionResult BearerAuthTest()
+    {
+        return Ok();
+    }
+
+    [HttpGet("[action]")]
+    public IActionResult ExceptionTest()
+    {
+        throw new Exception();
     }
 }