+++ src/WeihanLi.Web.Extensions/DataProtection/DataProtectionBuilderExtensions.cs
@@ -3,6 +3,7 @@
 
 using Microsoft.AspNetCore.DataProtection;
 using Microsoft.AspNetCore.Mvc;
+using WeihanLi.Common;
 using WeihanLi.Web.DataProtection.ParamsProtection;
 
 namespace WeihanLi.Web.DataProtection;
@@ -16,10 +17,7 @@ public static class DataProtectionBuilderExtensions
     /// <returns></returns>
     public static IDataProtectionBuilder AddParamsProtection(this IDataProtectionBuilder builder)
     {
-        if (null == builder)
-        {
-            throw new ArgumentNullException(nameof(builder));
-        }
+        Guard.NotNull(builder);
 
         builder.Services.Configure<MvcOptions>(action =>
             {
@@ -47,9 +45,8 @@ public static class DataProtectionBuilderExtensions
             throw new ArgumentNullException(nameof(optionsAction));
         }
 
-        builder.AddParamsProtection();
         builder.Services.Configure(optionsAction);
-
+        builder.AddParamsProtection();
         return builder;
     }
 }