+++ docs/ReleaseNotes.md
@@ -1,5 +1,20 @@
 # Release Notes
 
+## 1.3.0
+
+- Add `ApiResultFilter`
+
+## 1.2.1
+
+- add FluentAspectServiceProviderFactory
+- Add FeatureFlag extensions
+- Update HealthCheckExtension
+- Add Pager/DataProtection/AccessControlHelper
+- Add HttpContextTenantProvider
+- Add ApiKeyAuthentication Support, fixes Add API-Key based authentication #15
+- Add AuthorizationFilterAttribute
+- Implement JwtTokenService, fixes Simplify JWT usage #16
+
 ## 1.1.0
 
 - Header/Quey Auth